pyproject.toml
[tool.poetry]
name = "go-kafka-example"
version = "0.1.0"
description = ""
authors = ["Hank.Kuo <asdf024681029@gmail.com>"]
readme = "README.md"
packages = [{include = "go_kafka_example"}]

[tool.poetry.dependencies]
python = "^3.9"
kafka-python = "^2.0.2"
python-dotenv = "^1.0.0"
pyotp = "^2.9.0"
confluent-kafka = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
